Renamed error schema submodule to message
